✨ TWEAK: modular utilities
